refactor(api): move oidc authentication service registry
from lib to src/authentication
